requirements.txt
mistralai>=0.0.7
langgraph>=0.0.10
crewai>=0.1.0
chromadb>=0.4.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
ortools>=9.7.0
streamlit>=1.30.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0 
pydantic-settings
nest_asyncio>=0.1.0
plotly>=5.18.0
dateparser>=1.2.0
